Add files via upload
Added functions to display all values relevant to both matrices, A and B. 
Display optimal mixed strategies for both players. These stem from the values of alpha and beta.
Found the value of the game when beta,alpha in (0,1), which is given using the determinant and diagonal of each matrix.
I aim to check for a solution in pure strategies, called an equilibrium point, should one(or more) exist. These are both a column max in A and a row max in B, e.g. (1,1).