Add files via upload
The pure strategies functionality has been added for a 2x2 matrix, giving us a solution as an equilibrium point. These provide stability - where neither player wishes to change their strategy, and payoffs for each player which correspond to that specific entry in each matrix, e.g. (1,2) gives a_12 and b_12 as the values of the game for players 1 and 2 respectively. Now, we have issues for the degenerative cases in mixed strategies stemming from the graphs. The graphical solution in these cases may not be unique, which causes issues with displaying the strategies and payoffs for at least one of the players. The graph does however coincide with the solution in pure strategies in the examples I have tested thus far. Therefore, I will aim to deduce whether this is mathematically correct and adapt the code accordingly.